+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * `Vector2<T>` types with `Vec2` and `IVec2` aliases
 * Operators `+`, `+=`, `-`, `-=`, `*`, `*=`, `/` and `/=`
 * Support `no_std` (by disabling the `std` feature flag)
-* `approx_v05` feature: Implementation of [approx v0.5](https://docs.rs/approx/0.5) traits
+* `approx_v05` feature: Implementations of [approx v0.5](https://docs.rs/approx/0.5) traits
+* `serde` feature: Implementations of `Serialize` and `Deserialize` from [serde](https://docs.rs/serde/1)
 
 
 [Unreleased]: https://github.com/jcornaz/aline/compare/...HEAD